@@ -1,7 +1,7 @@
 import React from "react";
 
-import Card from "@material-ui/core/Card";
-import CardHeader from "@material-ui/core/CardHeader";
+import Card from "@mui/material/Card";
+import CardHeader from "@mui/material/CardHeader";
 
 import { TodoItem } from "./TodoItem";
 
