@@ -1,8 +1,8 @@
-import { createMuiTheme, responsiveFontSizes } from "@material-ui/core/styles";
-import { red } from "@material-ui/core/colors";
+import { createTheme, responsiveFontSizes, adaptV4Theme } from "@mui/material/styles";
+import { red } from "@mui/material/colors";
 
 // Create a theme instance.
-let theme = createMuiTheme({
+let theme = createTheme(adaptV4Theme({
   palette: {
     primary: {
       main: "#60afdf",
@@ -17,7 +17,7 @@ let theme = createMuiTheme({
       default: "#fff",
     },
   },
-});
+}));
 
 theme = responsiveFontSizes(theme);
 
